refactor(db): Split nac.rs into directory module
