+++ crates/db/src/nac/mod.rs
@@ -16,17 +16,21 @@
 //! - `acp.node_enable` config option (default: false)
 //! - Owner identity from the node's keyring
 
+pub mod factory;
+#[cfg(test)]
+mod tests;
+mod trait_impl;
+
+pub use factory::create_memory_nac_manager;
+#[cfg(not(target_arch = "wasm32"))]
+pub use factory::create_persistent_nac_manager;
+
 use std::sync::Arc;
 
 use acp::nac::{NacStatus, NodeACP, NodePermission};
-use acp::{MemoryZanzibarStore, ZanzibarStore};
+use acp::ZanzibarStore;
 use async_trait::async_trait;
 use identity::Did;
-
-#[cfg(not(target_arch = "wasm32"))]
-use acp::PersistentZanzibarStore;
-#[cfg(not(target_arch = "wasm32"))]
-use storage::RedbStore;
 
 use crate::error::{Error, Result};
 
@@ -126,19 +130,16 @@ impl<S: ZanzibarStore> NacManager<S> {
     /// This should be called during node startup. If NAC is configured to be
     /// enabled and an owner identity is provided, NAC will be activated.
     pub async fn initialize(&self, owner_identity: Option<&Did>) -> Result<()> {
-        // First, try to load existing NAC state
         self.nac
             .load()
             .await
             .map_err(|e| Error::Other(format!("failed to load NAC state: {}", e)))?;
 
-        // If NAC is configured to be enabled and we have an owner identity
         if self.config.enabled {
             let current_status = self.nac.status().await;
 
             match current_status {
                 NacStatus::NotConfigured => {
-                    // NAC needs to be enabled for the first time
                     if let Some(owner) = owner_identity {
                         self.nac
                             .enable(owner)
@@ -164,7 +165,6 @@ impl<S: ZanzibarStore> NacManager<S> {
                     );
                 }
                 NacStatus::DisabledTemporarily => {
-                    // Re-enable NAC since config says it should be enabled
                     self.nac
                         .re_enable()
                         .await
@@ -191,19 +191,11 @@ impl<S: ZanzibarStore> NacManager<S> {
     }
 
     /// Check if NAC is effectively enabled.
-    ///
-    /// NAC is effectively enabled when:
-    /// - Config says enabled AND
-    /// - NAC status is Enabled
     pub async fn is_enabled(&self) -> bool {
         self.config.enabled && self.nac.status().await == NacStatus::Enabled
     }
 
     /// Check if an identity has a specific node permission.
-    ///
-    /// Returns `true` if:
-    /// - NAC is not enabled (all operations allowed)
-    /// - The identity has the required permission
     pub async fn check_permission(
         &self,
         identity: &Did,
@@ -224,10 +216,6 @@ impl<S: ZanzibarStore> NacManager<S> {
     }
 
     /// Check if an identity is an admin based on stored relationships.
-    ///
-    /// Unlike `is_admin()`, this checks actual stored relationships regardless
-    /// of NAC status. Used for operations like re-enable where we verify admin
-    /// access even when NAC is temporarily disabled.
     pub async fn is_admin_persisted(&self, identity: &Did) -> Result<bool> {
         self.nac
             .is_admin_persisted(identity)
@@ -241,8 +229,6 @@ impl<S: ZanzibarStore> NacManager<S> {
     }
 
     /// Enable NAC with the given owner.
-    ///
-    /// This can only be called when NAC is not already enabled.
     pub async fn enable(&self, owner: &Did) -> Result<()> {
         self.nac
             .enable(owner)
@@ -251,10 +237,7 @@ impl<S: ZanzibarStore> NacManager<S> {
     }
 
     /// Temporarily disable NAC.
-    ///
-    /// The requestor must be an admin.
     pub async fn disable(&self, requestor: &Did) -> Result<()> {
-        // Only admins can disable
         if !self.is_admin(requestor).await? {
             return Err(Error::Acp("only admins can disable NAC".into()));
         }
@@ -266,11 +249,7 @@ impl<S: ZanzibarStore> NacManager<S> {
     }
 
     /// Re-enable NAC after temporary disable.
-    ///
-    /// The requestor must be an admin. Uses persisted admin check since
-    /// NAC is disabled (is_admin returns true for everyone when disabled).
     pub async fn re_enable(&self, requestor: &Did) -> Result<()> {
-        // Use persisted check since is_admin returns true for everyone when disabled
         if !self.is_admin_persisted(requestor).await? {
             return Err(Error::Acp("only admins can re-enable NAC".into()));
         }
@@ -282,14 +261,11 @@ impl<S: ZanzibarStore> NacManager<S> {
     }
 
     /// Purge all NAC state (dev mode only).
-    ///
-    /// The requestor must be an admin and dev mode must be enabled.
     pub async fn purge(&self, requestor: &Did) -> Result<()> {
         if !self.config.dev_mode {
             return Err(Error::Acp("NAC purge is only allowed in dev mode".into()));
         }
 
-        // Only admins can purge
         if !self.is_admin(requestor).await? {
             return Err(Error::Acp("only admins can purge NAC".into()));
         }
@@ -301,8 +277,6 @@ impl<S: ZanzibarStore> NacManager<S> {
     }
 
     /// Add an admin relationship.
-    ///
-    /// The requestor must be an admin.
     pub async fn add_admin(&self, requestor: &Did, target: &Did) -> Result<bool> {
         self.nac
             .add_admin(requestor, target)
@@ -311,8 +285,6 @@ impl<S: ZanzibarStore> NacManager<S> {
     }
 
     /// Remove an admin relationship.
-    ///
-    /// The requestor must be an admin. The owner cannot be removed.
     pub async fn remove_admin(&self, requestor: &Did, target: &Did) -> Result<bool> {
         self.nac
             .remove_admin(requestor, target)
@@ -350,33 +322,19 @@ impl<S: ZanzibarStore> NacManager<S> {
     pub fn config(&self) -> &NacConfig {
         &self.config
     }
-}
 
-/// Create an in-memory NAC manager for testing.
-pub fn create_memory_nac_manager(config: NacConfig) -> NacManager<MemoryZanzibarStore> {
-    let store = Arc::new(MemoryZanzibarStore::new());
-    NacManager::new(store, config)
-}
+    /// Get NAC info for HTTP responses.
+    pub async fn info(&self) -> NacInfo {
+        let status = self.status().await;
+        let owner = self.owner().await;
 
-/// Create a persistent NAC manager.
-///
-/// The NAC data is stored in a separate directory (`local_node_acp/`) under the data path.
-#[cfg(not(target_arch = "wasm32"))]
-pub fn create_persistent_nac_manager(
-    data_path: &std::path::Path,
-) -> Result<NacManager<PersistentZanzibarStore<RedbStore>>> {
-    let nac_path = data_path.join("local_node_acp");
-    std::fs::create_dir_all(&nac_path)
-        .map_err(|e| Error::Other(format!("failed to create NAC data directory: {}", e)))?;
-
-    let db_path = nac_path.join("nac.db");
-    let store = PersistentZanzibarStore::open(&db_path)
-        .map_err(|e| Error::Acp(format!("failed to open NAC store: {}", e)))?;
-
-    Ok(NacManager::new(
-        Arc::new(store),
-        NacConfig::default().with_data_path(nac_path.display().to_string()),
-    ))
+        NacInfo {
+            status: status.to_string(),
+            configured_enabled: self.config.enabled,
+            dev_mode: self.config.dev_mode,
+            owner: owner.map(|d| d.to_string()),
+        }
+    }
 }
 
 /// Information about NAC status for HTTP responses.
@@ -394,229 +352,4 @@ pub struct NacInfo {
     /// Owner DID if NAC is enabled
     #[serde(skip_serializing_if = "Option::is_none")]
     pub owner: Option<String>,
-}
-
-impl<S: ZanzibarStore> NacManager<S> {
-    /// Get NAC info for HTTP responses.
-    pub async fn info(&self) -> NacInfo {
-        let status = self.status().await;
-        let owner = self.owner().await;
-
-        NacInfo {
-            status: status.to_string(),
-            configured_enabled: self.config.enabled,
-            dev_mode: self.config.dev_mode,
-            owner: owner.map(|d| d.to_string()),
-        }
-    }
-}
-
-#[cfg_attr(not(target_arch = "wasm32"), async_trait)]
-#[cfg_attr(target_arch = "wasm32", async_trait(?Send))]
-impl<S: ZanzibarStore> NacManagerApi for NacManager<S> {
-    async fn initialize(&self, owner_identity: Option<&Did>) -> Result<()> {
-        NacManager::initialize(self, owner_identity).await
-    }
-    async fn status(&self) -> NacStatus {
-        NacManager::status(self).await
-    }
-    async fn owner(&self) -> Option<Did> {
-        NacManager::owner(self).await
-    }
-    async fn is_enabled(&self) -> bool {
-        NacManager::is_enabled(self).await
-    }
-    async fn check_permission(&self, identity: &Did, permission: NodePermission) -> Result<bool> {
-        NacManager::check_permission(self, identity, permission).await
-    }
-    async fn is_admin(&self, identity: &Did) -> Result<bool> {
-        NacManager::is_admin(self, identity).await
-    }
-    async fn is_admin_persisted(&self, identity: &Did) -> Result<bool> {
-        NacManager::is_admin_persisted(self, identity).await
-    }
-    async fn is_owner(&self, identity: &Did) -> bool {
-        NacManager::is_owner(self, identity).await
-    }
-    async fn enable(&self, owner: &Did) -> Result<()> {
-        NacManager::enable(self, owner).await
-    }
-    async fn disable(&self, requestor: &Did) -> Result<()> {
-        NacManager::disable(self, requestor).await
-    }
-    async fn re_enable(&self, requestor: &Did) -> Result<()> {
-        NacManager::re_enable(self, requestor).await
-    }
-    async fn purge(&self, requestor: &Did) -> Result<()> {
-        NacManager::purge(self, requestor).await
-    }
-    async fn add_admin(&self, requestor: &Did, target: &Did) -> Result<bool> {
-        NacManager::add_admin(self, requestor, target).await
-    }
-    async fn remove_admin(&self, requestor: &Did, target: &Did) -> Result<bool> {
-        NacManager::remove_admin(self, requestor, target).await
-    }
-    async fn add_permission_grant(
-        &self,
-        requestor: &Did,
-        target: &Did,
-        permission: NodePermission,
-    ) -> Result<bool> {
-        NacManager::add_permission_grant(self, requestor, target, permission).await
-    }
-    async fn remove_permission_grant(
-        &self,
-        requestor: &Did,
-        target: &Did,
-        permission: NodePermission,
-    ) -> Result<bool> {
-        NacManager::remove_permission_grant(self, requestor, target, permission).await
-    }
-    async fn info(&self) -> NacInfo {
-        NacManager::info(self).await
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    fn test_did() -> Did {
-        Did::new("did:key:z6MkhaXgBZDvotDkL5257faiztiGiC2QtKLGpbnnEGta2doK").unwrap()
-    }
-
-    fn test_did2() -> Did {
-        Did::new("did:key:z6MkpTHR8VNsBxYAAWHut2Geadd9jSwuBV8xRoAnwWsdvktH").unwrap()
-    }
-
-    #[tokio::test]
-    async fn test_nac_manager_disabled_by_default() {
-        let manager = create_memory_nac_manager(NacConfig::default());
-
-        assert!(!manager.is_enabled().await);
-        assert_eq!(manager.status().await, NacStatus::NotConfigured);
-
-        // All permissions should be allowed when disabled
-        let identity = test_did();
-        let allowed = manager
-            .check_permission(&identity, NodePermission::DacBypass)
-            .await
-            .unwrap();
-        assert!(allowed);
-    }
-
-    #[tokio::test]
-    async fn test_nac_manager_enable() {
-        let manager = create_memory_nac_manager(NacConfig::new().with_enabled());
-
-        let owner = test_did();
-        manager.initialize(Some(&owner)).await.unwrap();
-
-        assert!(manager.is_enabled().await);
-        assert_eq!(manager.status().await, NacStatus::Enabled);
-        assert!(manager.is_owner(&owner).await);
-    }
-
-    #[tokio::test]
-    async fn test_nac_manager_permission_check() {
-        let manager = create_memory_nac_manager(NacConfig::new().with_enabled());
-
-        let owner = test_did();
-        let other = test_did2();
-        manager.initialize(Some(&owner)).await.unwrap();
-
-        // Owner has all permissions
-        assert!(manager
-            .check_permission(&owner, NodePermission::DacBypass)
-            .await
-            .unwrap());
-
-        // Non-owner denied
-        assert!(!manager
-            .check_permission(&other, NodePermission::DacBypass)
-            .await
-            .unwrap());
-    }
-
-    #[tokio::test]
-    async fn test_nac_manager_admin() {
-        let manager = create_memory_nac_manager(NacConfig::new().with_enabled());
-
-        let owner = test_did();
-        let admin = test_did2();
-        manager.initialize(Some(&owner)).await.unwrap();
-
-        // Add admin
-        let added = manager.add_admin(&owner, &admin).await.unwrap();
-        assert!(added);
-
-        // Admin has all permissions
-        assert!(manager
-            .check_permission(&admin, NodePermission::DacBypass)
-            .await
-            .unwrap());
-    }
-
-    #[tokio::test]
-    async fn test_nac_manager_disable_reenable() {
-        let manager = create_memory_nac_manager(NacConfig::new().with_enabled());
-
-        let owner = test_did();
-        let other = test_did2();
-        manager.initialize(Some(&owner)).await.unwrap();
-
-        // Disable
-        manager.disable(&owner).await.unwrap();
-        assert_eq!(manager.status().await, NacStatus::DisabledTemporarily);
-
-        // All operations allowed when disabled
-        assert!(manager
-            .check_permission(&other, NodePermission::DacBypass)
-            .await
-            .unwrap());
-
-        // Re-enable
-        manager.re_enable(&owner).await.unwrap();
-        assert_eq!(manager.status().await, NacStatus::Enabled);
-
-        // Non-owner denied again
-        assert!(!manager
-            .check_permission(&other, NodePermission::DacBypass)
-            .await
-            .unwrap());
-    }
-
-    #[tokio::test]
-    async fn test_nac_manager_purge_requires_dev_mode() {
-        let manager = create_memory_nac_manager(NacConfig::new().with_enabled());
-
-        let owner = test_did();
-        manager.initialize(Some(&owner)).await.unwrap();
-
-        // Purge should fail without dev mode
-        let result = manager.purge(&owner).await;
-        assert!(result.is_err());
-
-        // With dev mode
-        let manager = create_memory_nac_manager(NacConfig::new().with_enabled().with_dev_mode());
-        manager.initialize(Some(&owner)).await.unwrap();
-
-        let result = manager.purge(&owner).await;
-        assert!(result.is_ok());
-        assert_eq!(manager.status().await, NacStatus::NotConfigured);
-    }
-
-    #[tokio::test]
-    async fn test_nac_manager_info() {
-        let manager = create_memory_nac_manager(NacConfig::new().with_enabled().with_dev_mode());
-
-        let owner = test_did();
-        manager.initialize(Some(&owner)).await.unwrap();
-
-        let info = manager.info().await;
-        assert_eq!(info.status, "enabled");
-        assert!(info.configured_enabled);
-        assert!(info.dev_mode);
-        assert!(info.owner.is_some());
-    }
 }
